WindowsInstaller now doesn't compare a string using != AddSVN now works with mawk aswell as gawk
git-svn-id: http://svn.dmdirc.com/trunk@3627 00569f92-eb28-0410-84fd-f71c24880f43

+++ AddSVN.sh
@@ -1,12 +1,33 @@
 #!/bin/sh
 
 SVN=`which svn`
+AWK=`which gawk`
+TYPE="gawk"
+AWK=""
+if [ "" = "${AWK}" ]; then
+	AWK=`which mawk`
+	TYPE="mawk"
+	if [ "" = "${AWK}" ]; then
+		echo "Unknown awk variation, not running."
+		exit 0;
+	fi;
+fi;
 
 SVNREV=`$SVN info | grep Revision`
 SVNREV=${SVNREV##*: }
 
-PRE='int SVN_REVISION = '
-POST='int SVN_REVISION = '${SVNREV}'; \/\/ '
+if [ "${TYPE}" = "mawk" ]; then
+	PRE='int SVN_REVISION = '
+	POST='int SVN_REVISION = '${SVNREV}';'
+	if [ "${1}" = "--pre" ]; then
+		POST=${POST}' // ';
+	elif [ "${1}" = "--post" ]; then
+		POST=${POST}' \/\/ ';
+	fi;
+elif [ "${TYPE}" = "gawk" ]; then
+	PRE='int SVN_REVISION = '
+	POST='int SVN_REVISION = '${SVNREV}'; \/\/ '
+fi;
 
 OLD=""
 
@@ -21,6 +42,9 @@ elif [ "${1}" = "--post" ]; then
 fi;
 
 if [ "" != "${OLD}" ]; then
-	awk '{gsub(/'"${OLD}"'/,"'"${NEW}"'");print}' ${PWD}/src/com/dmdirc/Main.java > ${PWD}/src/com/dmdirc/Main.java.tmp 2>/dev/null
-	mv ${PWD}/src/com/dmdirc/Main.java.tmp ${PWD}/src/com/dmdirc/Main.java
+	${AWK} '{gsub(/'"${OLD}"'/,"'"${NEW}"'");print}' ${PWD}/src/com/dmdirc/Main.java > ${PWD}/src/com/dmdirc/Main.java.tmp 2>/dev/null
+	if [ -e ${PWD}/src/com/dmdirc/Main.java.tmp ]; then
+		mv ${PWD}/src/com/dmdirc/Main.java.tmp ${PWD}/src/com/dmdirc/Main.java
+	fi;
+	# cat ${PWD}/src/com/dmdirc/Main.java | grep SVN
 fi;
